% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bestfitr-package.R
\docType{package}
\name{bestfitr-package}
\alias{bestfitr}
\alias{bestfitr-package}
\title{bestfitr: Bayesian Flood-Frequency and Extreme-Value Analysis}
\description{
R bindings to a shared C++ port of the USACE-RMC Numerics and RMC.BestFit libraries for flood-frequency and extreme-value analysis. This early version exposes the Generalized Extreme Value distribution: density, distribution and quantile functions, moments, parameter estimation (method of moments, L-moments, maximum likelihood), and quantile uncertainty.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://cameronbracken.github.io/bestfit/r/}
  \item \url{https://github.com/cameronbracken/bestfit}
  \item Report bugs at \url{https://github.com/cameronbracken/bestfit/issues}
}

}
\author{
\strong{Maintainer}: Cam Bracken \email{cameron.bracken@pm.me}

Authors:
\itemize{
  \item Cam Bracken \email{cameron.bracken@pm.me}
}

}
\keyword{internal}
